requirements.txt
djangorestframework-simplejwt
django-filter
asgiref==3.10.0
Django==5.2.7
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
phonenumbers==9.0.17
sqlparse==0.5.3
